# database/config/docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:7001",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/runtime/save": {
      "post": {
        "tags": [
          "runtime-controller"
        ],
        "summary": "Save a runtime object",
        "description": "Saves a new runtime and returns the saved runtime.",
        "operationId": "save_2",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuntimeDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeDto"
                }
              }
            }
          }
        }
      }
    },
    "/runtime/save-all": {
      "post": {
        "tags": [
          "runtime-controller"
        ],
        "summary": "Save multiple runtimes",
        "description": "Saves a list of runtimes and returns the list of saved runtimes.",
        "operationId": "save_3",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/RuntimeDto"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RuntimeDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/history/save": {
      "post": {
        "tags": [
          "history-controller"
        ],
        "summary": "Save a history",
        "description": "Saves a new history and returns the saved history.",
        "operationId": "save_4",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HistoryDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryDto"
                }
              }
            }
          }
        }
      }
    },
    "/history/save-all": {
      "post": {
        "tags": [
          "history-controller"
        ],
        "summary": "Save multiple histories",
        "description": "Saves a list of histories and returns the list of saved histories.",
        "operationId": "save_5",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/HistoryDto"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HistoryDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/runtime/get/{id}": {
      "get": {
        "tags": [
          "runtime-controller"
        ],
        "summary": "Retrieve a runtime by its ID",
        "description": "Retrieves a runtime based on its ID. Yields HttpStatus.NOT_FOUND if no runtime was found.",
        "operationId": "getRuntimeById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeDto"
                }
              }
            }
          }
        }
      }
    },
    "/runtime/get-for-drawing/{id}": {
      "get": {
        "tags": [
          "runtime-controller"
        ],
        "summary": "Retrieve runtimes by their associated drawing ID",
        "description": "Retrieves a list of all runtimes associated with a specific drawing ID.Yields HttpStatus.NOT_FOUND if no runtime was found.",
        "operationId": "getRuntimesByDrawingId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RuntimeDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/runtime/get-all": {
      "get": {
        "tags": [
          "runtime-controller"
        ],
        "summary": "Retrieve all runtimes",
        "description": "Retrieves a list of all runtimes.Yields an empty list if no runtimes were found.",
        "operationId": "getAllRuntimes",
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RuntimeDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/history/get/{id}": {
      "get": {
        "tags": [
          "history-controller"
        ],
        "summary": "Retrieve a history by its ID",
        "description": "Retrieves a history based on its ID. Yields HttpStatus.NOT_FOUND if no history was found.",
        "operationId": "getHistoryById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryDto"
                }
              }
            }
          }
        }
      }
    },
    "/history/get-all": {
      "get": {
        "tags": [
          "history-controller"
        ],
        "summary": "Retrieve all histories",
        "description": "Retrieves a list of all histories. Yields an empty list if no histories were found.",
        "operationId": "getAllHistories",
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HistoryDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/runtime/delete/{id}": {
      "delete": {
        "tags": [
          "runtime-controller"
        ],
        "summary": "Delete a runtime by its ID",
        "description": "Deletes a runtime based on its ID.",
        "operationId": "deleteRuntimeById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/runtime/delete-for-drawing/{id}": {
      "delete": {
        "tags": [
          "runtime-controller"
        ],
        "summary": "Delete runtimes by ist associated drawing ID",
        "description": "Deletes all runtimes associated with a specific drawing ID.",
        "operationId": "deleteRuntimesByDrawingId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/history/delete/{id}": {
      "delete": {
        "tags": [
          "history-controller"
        ],
        "summary": "Delete a history by its ID",
        "description": "Deletes a history based on its ID.",
        "operationId": "deleteHistoryById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "RuntimeDto": {
        "type": "object",
        "properties": {
          "runtime_id": {
            "type": "integer",
            "format": "int32"
          },
          "drawing_id": {
            "type": "integer",
            "format": "int32"
          },
          "machine": {
            "type": "string"
          },
          "machine_runtime": {
            "type": "number",
            "format": "float"
          }
        }
      },
      "FeedbackDto": {
        "type": "object",
        "properties": {
          "history_id": {
            "type": "integer",
            "format": "int32"
          },
          "drawing_id": {
            "type": "integer",
            "format": "int32"
          },
          "feedback_desc": {
            "type": "string"
          },
          "feedback_value": {
            "type": "integer",
            "format": "int32"
          },
          "feedback_id": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          }
        },
        "readOnly": true
      },
      "HistoryDto": {
        "type": "object",
        "properties": {
          "query_drawing": {
            "type": "string"
          },
          "query_path": {
            "type": "string"
          },
          "history_id": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "feedbacks": {
            "type": "array",
            "readOnly": true,
            "items": {
              "$ref": "#/components/schemas/FeedbackDto"
            }
          }
        }
      },
...
    }
  }
}